refactor(world-state): remove client-side WorldStateOpsQueue
Read/write ordering is now enforced server-side by the wsdb per-fork
scheduler, so the IPC client no longer needs its own per-fork ordering
queue: every op is sent immediately and the server serializes writes per
fork while running reads concurrently. IpcWorldState.execute() now sends
directly; the per-fork queue map, its lifecycle (stop on close/deleteFork),
and the WorldStateOpsQueue class are deleted. The WorldStateOperationName
label type moves to world_state_operation.ts (still used for
instrumentation). The A-1055 delayed-close-fork regression test drops its
now-obsolete per-fork-queue-cleanup assertion (the silent-dispose check
remains).

native_world_state.test.ts (50 tests) passes against the async server,
including "Concurrent requests" — which, with the client queue gone, now
genuinely exercises server-side per-fork read/write ordering.
